Ajout d'une classe Packet et modification des indexs et méthodes de Packet,Sensor, et Server

+++ intermediaire/Sensor.hpp
@@ -18,9 +18,9 @@
 class Sensor
 {
 private:
-    float m_data; //la donnée aléatoire à fournir
-    float m_minData;
-    float m_maxData;
+    int m_data; //la donnée aléatoire à fournir
+    int m_minData;
+    int m_maxData;
 public:
 // Definition de la forme canonique
     Sensor():m_data(), m_minData(), m_maxData(){}; //Constructeur pas défaut
@@ -28,8 +28,8 @@ public:
     Sensor& operator=(const Sensor& param_s); //Opérateur d'affectation
     ~Sensor(); //Destructeur
 
-    float aleaGenVal(); //Nous pouvons déclarer le type dans la classe sensor pour le rendu intermédiaire
-    float getData();
+    int aleaGenVal(); //Nous pouvons déclarer le type dans la classe sensor pour le rendu intermédiaire
+    int getData();
 };
